+++ tests/e2e/snapshots/support/stub-websocket.ts
@@ -1,0 +1,51 @@
+import type { Page } from "@playwright/test";
+
+/**
+ * Stubs the browser's native `WebSocket` so conversation-page snapshot
+ * tests don't trigger real connection attempts (which fail in CI with
+ * "Failed to connect to server" toasts because no agent-server is
+ * running at the Vite proxy target).
+ *
+ * Must be called **before** `page.goto()` — it uses `page.addInitScript`
+ * so the stub is in place before any app code runs.
+ */
+export async function stubWebSocket(page: Page): Promise<void> {
+  await page.addInitScript(() => {
+    const noop = () => {};
+    class StubWebSocket extends EventTarget {
+      static CONNECTING = 0;
+      static OPEN = 1;
+      static CLOSING = 2;
+      static CLOSED = 3;
+      readyState = StubWebSocket.OPEN;
+      url: string;
+      protocol = "";
+      extensions = "";
+      bufferedAmount = 0;
+      binaryType: BinaryType = "blob";
+      onopen: ((ev: Event) => void) | null = null;
+      onclose: ((ev: CloseEvent) => void) | null = null;
+      onmessage: ((ev: MessageEvent) => void) | null = null;
+      onerror: ((ev: Event) => void) | null = null;
+
+      constructor(url: string | URL) {
+        super();
+        this.url = typeof url === "string" ? url : url.toString();
+        setTimeout(() => {
+          const evt = new Event("open");
+          this.onopen?.(evt);
+          this.dispatchEvent(evt);
+        }, 0);
+      }
+
+      send = noop;
+      close = noop;
+      CONNECTING = StubWebSocket.CONNECTING;
+      OPEN = StubWebSocket.OPEN;
+      CLOSING = StubWebSocket.CLOSING;
+      CLOSED = StubWebSocket.CLOSED;
+    }
+    (window as unknown as { WebSocket: unknown }).WebSocket =
+      StubWebSocket as unknown as typeof WebSocket;
+  });
+}
